#61 / Rename response data class
