+++ app/src/main/java/sopt/uni/data/source/remote/response/ResponseShortGameResultDto.kt
@@ -1,10 +1,11 @@
-package sopt.uni.data.entity.shortgame
+package sopt.uni.data.source.remote.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
+import sopt.uni.data.entity.shortgame.RoundMission
 
 @Serializable
-data class ResponseShortGameResult(
+data class ResponseShortGameResultDto(
     @SerialName("myRoundMission")
     val myRoundMission: RoundMission,
     @SerialName("partnerRoundMission")
